+++ packages/fluentui/react-northstar/src/components/Datepicker/Datepicker.tsx
@@ -111,7 +111,6 @@ export const Datepicker: ComponentWithAs<'div', DatepickerProps> &
   );
 
   const { calendar, popup, input, className, design, styles, variables, formatMonthDayYear } = props;
-  const valueFormatter = date => (date ? formatMonthDayYear(date) : '');
 
   const nonNullSelectedDate = selectedDate ?? props.today ?? new Date();
 
@@ -144,6 +143,7 @@ export const Datepicker: ComponentWithAs<'div', DatepickerProps> &
     invalidInputErrorMessage: props.invalidInputErrorMessage,
     isOutOfBoundsErrorMessage: props.isOutOfBoundsErrorMessage,
     goToToday: props.goToToday,
+    openCalendarTitle: props.openCalendarTitle,
     prevMonthAriaLabel: props.prevMonthAriaLabel,
     nextMonthAriaLabel: props.nextMonthAriaLabel,
     prevYearAriaLabel: props.prevYearAriaLabel,
@@ -157,6 +157,8 @@ export const Datepicker: ComponentWithAs<'div', DatepickerProps> &
     selectedDateFormatString: props.selectedDateFormatString,
     todayDateFormatString: props.todayDateFormatString,
   };
+
+  const valueFormatter = date => (date ? formatMonthDayYear(date, dateFormatting) : '');
 
   const ElementType = getElementType(props);
   const unhandledProps = useUnhandledProps(Datepicker.handledProps, props);
@@ -321,6 +323,7 @@ Datepicker.propTypes = {
   invalidInputErrorMessage: PropTypes.string,
   isOutOfBoundsErrorMessage: PropTypes.string,
   goToToday: PropTypes.string,
+  openCalendarTitle: PropTypes.string,
   prevMonthAriaLabel: PropTypes.string,
   nextMonthAriaLabel: PropTypes.string,
   prevYearAriaLabel: PropTypes.string,
@@ -349,11 +352,6 @@ Datepicker.defaultProps = {
   required: false,
 
   ...DEFAULT_CALENDAR_STRINGS,
-
-  // TODO: move defaults to date-time-utilities
-  isRequiredErrorMessage: 'A date selection is required',
-  invalidInputErrorMessage: 'Manually entered date is not in correct format.',
-  isOutOfBoundsErrorMessage: 'The selected date is from the restricted range.',
 };
 
 Datepicker.handledProps = Object.keys(Datepicker.propTypes) as any;