+++ packages/fluentui/react-northstar/src/themes/teams/components/Skeleton/skeletonVariables.ts
@@ -6,6 +6,7 @@ export interface SkeletonVariables {
   lineBackground?: string;
   shapeBackground?: string;
   animationBackground?: string;
+  animationBackgroundSecondary?: string;
 
   // Skeleton Button
   buttonHeight?: string;
@@ -19,6 +20,16 @@ export interface SkeletonVariables {
   inputHeight?: string;
   inputWidth?: string;
   inputBackground?: string;
+
+  // Skeleton Avatar
+  avatarBackground?: string;
+  avatarSmallest?: string;
+  avatarSmaller?: string;
+  avatarSmall?: string;
+  avatarMedium?: string;
+  avatarLarge?: string;
+  avatarLarger?: string;
+  avatarLargest?: string;
 }
 
 export const skeletonVariables = (siteVariables): SkeletonVariables => {
@@ -26,6 +37,7 @@ export const skeletonVariables = (siteVariables): SkeletonVariables => {
     lineBackground: siteVariables.colorScheme.default.background4,
     shapeBackground: siteVariables.colorScheme.default.background4,
     animationBackground: siteVariables.colorScheme.default.background,
+    animationBackgroundSecondary: siteVariables.colorScheme.default.background5,
     lineMargin: `0 0 ${pxToRem(2)} 0`,
     shapeMargin: `0 0 ${pxToRem(2)} 0`,
 
@@ -41,5 +53,15 @@ export const skeletonVariables = (siteVariables): SkeletonVariables => {
     inputHeight: pxToRem(32),
     inputWidth: pxToRem(154),
     inputBackground: siteVariables.colorScheme.default.background4,
+
+    // Skeleton Avatar
+    avatarBackground: siteVariables.colorScheme.default.background4,
+    avatarSmallest: pxToRem(20),
+    avatarSmaller: pxToRem(24),
+    avatarSmall: pxToRem(28),
+    avatarMedium: pxToRem(32),
+    avatarLarge: pxToRem(44),
+    avatarLarger: pxToRem(64),
+    avatarLargest: pxToRem(96),
   };
 };